requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
opencv-python==4.8.1.78
numpy==1.24.3
scipy==1.11.4
pillow==10.1.0
python-socketio==5.10.0
eventlet==0.33.3
gunicorn==21.2.0



